requirements.txt
flask 
dotenv
mysql-connector-python